Fix Unet cropping for same padding
With `same` padding, the output size of a Unet should be equal to its input size.
The cropping function in `Upsample` that ensures translation equivariance
with the total upsampling factor for nets with `valid` padding breaks this
assumption for certain Unet instances.
However, a Unet instance with `same` padding cannot be translation-equivariant,
therefore the cropping mechanism should not be used for nets with `valid` padding.